requirements-dev.txt
tox==4.6.4
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.31.0
urllib3==2.0.2
web3
eth-accounts
py_ecc